// frontend/api/models/SpecializedUserCreateRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SpecializedUserCreateRequest = {
    username: string;
    password: string;
    realName: string;
    mobile?: string;
    institutionId?: string;
    regionCode?: string;
    status?: 'enabled' | 'disabled';
};